# Conflicts:
#	app/src/main/res/values/strings.xml
#	tronlib/src/main/java/com/devband/tronlib/TronNetwork.java

+++ app/src/main/java/com/devband/tronwalletforandroid/ui/accountdetail/transfer/TransferFragment.java
@@ -39,7 +39,7 @@ public class TransferFragment extends BaseFragment implements TransferView {
     private AdapterView mAdapterView;
     private AccountTransferAdapter mAdapter;
 
-    private int mStartIndex = 0;
+    private long mStartIndex = 0;
 
     private boolean mIsLoading;
