+++ tronlib/src/main/java/com/devband/tronlib/services/TronScanService.java
@@ -22,11 +22,11 @@ public interface TronScanService {
     Single<Transfers> getTransfers(@Query("address") String address, @Query("token") String token);
 
     @GET("api/transfer")
-    Single<Transfers> getTransfers(@Query("address") String address, @Query("start") int start,
-                                   @Query("limit") int limit, @Query("sort") String sort, @Query("count") boolean count);
+    Single<Transfers> getTransfers(@Query("address") String address, @Query("start") long start,
+            @Query("limit") int limit, @Query("sort") String sort, @Query("count") boolean count);
 
     @GET("api/transfer")
-    Single<Transfers> getTransfers(@Query("start") int start, @Query("limit") int limit,
+    Single<Transfers> getTransfers(@Query("start") long start, @Query("limit") int limit,
             @Query("sort") String sort, @Query("count") boolean count);
 
     @GET("api/transfer")
@@ -38,16 +38,16 @@ public interface TronScanService {
     Single<List<Market>> getMarket();
 
     @GET("api/block")
-    Single<Blocks> getBlock(@Query("sort") String sort, @Query("limit") int limit, @Query("start") int start);
+    Single<Blocks> getBlock(@Query("sort") String sort, @Query("limit") int limit, @Query("start") long start);
 
     @GET("api/account/{address}/stats")
     Single<TransactionStats> getTransactionStats(@Path("address") String address);
 
     @GET("api/transaction")
-    Single<Transactions> getTransactions(@Query("start") int start, @Query("limit") int limit,
+    Single<Transactions> getTransactions(@Query("start") long start, @Query("limit") int limit,
             @Query("sort") String sort, @Query("count") boolean count);
 
     @GET("api/transaction")
-    Single<Transactions> getTransactions(@Query("address") String address, @Query("start") int start, @Query("limit") int limit,
+    Single<Transactions> getTransactions(@Query("address") String address, @Query("start") long start, @Query("limit") int limit,
             @Query("sort") String sort, @Query("count") boolean count);
 }
